pyproject.toml
[tool.poetry]
name = "jawafdehi"
version = "0.1.0"
description = ""
authors = ["Damodar Dahal <damo94761@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.8"
djangorestframework = "^3.16.1"
django-cors-headers = "^4.9.0"
psycopg2-binary = "^2.9.11"
python-dotenv = "^1.2.1"
# nepal-entity-service = {git = "https://github.com/NewNepal-org/NepalEntityService.git", extras = ["api", "scraping"]}
nepal-entity-service = {path = "../NepalEntityService", extras = ["api", "scraping", "agni"], develop = true}
django-filter = "^25.2"
drf-spectacular = "^0.29.0"
dj-database-url = "^3.0.1"
django-jazzmin = "^3.0.1"
django-tinymce = "^5.0.0"
rules = "^3.5"
whitenoise = "^6.11.0"
django-auditlog = "^3.3.0"
gunicorn = "^23.0.0"
django-q2 = "^1.7.4"
jsonref = "^1.1.0"


[tool.poetry.group.dev.dependencies]
hypothesis = "^6.148.2"
pytest = "^9.0.1"
pytest-django = "^4.11.1"
greenlet = "^3.3.0"
gevent = "^25.9.1"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--reuse-db"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
